feat(get-page-url.use-case.ts): add support for custom endpoint in GetPageUrlUseCase
This commit adds support for a custom endpoint in the GetPageUrlUseCase. The custom endpoint can be passed as a parameter to the execute method of the use case. The endpoint must contain the placeholders '{steamID64}', '{appID}', and '{contextID}' and must include the 'http://' or 'https://' protocol. If any of these conditions are not met, a UseCaseException is thrown.