feat(http-client): add http-client module with get method and tests
The http-client module was added to the project to handle HTTP requests. It
contains a get method that receives an object with the URL and headers of the
request and returns a Promise with the response data. The module also has a
setDefaultHeaders method to set default headers for all requests and a
setDefaultCookies method to set default cookies for all requests.

The module was tested with a unit test that checks if the get method sends the
request with the correct headers. The test uses the fastify and find-open-port
libraries to create a local server and find an available port to listen to.